feat: 이미지 변경 훅 커스텀 훅으로 변경
