+++ apps/web/src/widgets/setting/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useProfileImageEditDialog } from './useProfileImageEditDialog';
