+++ apps/web/src/widgets/setting/model/index.ts
@@ -3,3 +3,4 @@ export * from './navigation';
 export * from './activity-mode';
 export * from './tabs';
 export * from './setting-navigation';
+export * from './hooks';
