+++ apps/web/src/widgets/setting/ui/profile-image-edit-dialog/ProfileImageEditDialog.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { useEffect, useMemo, useRef, useState } from 'react';
-
 import {
   CTAButton,
   Dialog,
@@ -14,11 +12,10 @@ import {
 } from '@causw/cds';
 
 import { DEFAULT_PROFILE_IMAGE_TYPES } from '@/shared/constants';
-import type {
-  ProfileImageEditValue,
-  UserProfileImageType,
-} from '@/shared/types';
+import type { ProfileImageEditValue } from '@/shared/types';
 import { ProfileAvatar } from '@/shared/ui';
+
+import { useProfileImageEditDialog } from '../../model/hooks';
 
 export interface ProfileImageEditDialogProps {
   open: boolean;
@@ -35,91 +32,24 @@ export const ProfileImageEditDialog = ({
   onSubmit,
   requireSubmitToClose = false,
 }: ProfileImageEditDialogProps) => {
-  const fileInputRef = useRef<HTMLInputElement | null>(null);
-  const objectUrlRef = useRef<string | null>(null);
-  const allowCloseRef = useRef(false);
-
-  const [selectedType, setSelectedType] = useState<UserProfileImageType>(
-    initialValue.profileImageType,
-  );
-  const [customImageUrl, setCustomImageUrl] = useState<string | null>(
-    initialValue.profileImageType === 'CUSTOM'
-      ? (initialValue.profileImageUrl ?? null)
-      : null,
-  );
-  const [customImageFile, setCustomImageFile] = useState<File | null>(
-    initialValue.customImageFile ?? null,
-  );
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  useEffect(() => {
-    return () => {
-      if (objectUrlRef.current) {
-        URL.revokeObjectURL(objectUrlRef.current);
-      }
-    };
-  }, []);
-
-  const previewUrl = selectedType === 'CUSTOM' ? customImageUrl : null;
-  const isSubmitDisabled =
-    selectedType === initialValue.profileImageType &&
-    (selectedType !== 'CUSTOM'
-      ? true
-      : (customImageUrl ?? null) === (initialValue.profileImageUrl ?? null));
-
-  const selectedValue = useMemo<ProfileImageEditValue>(
-    () => ({
-      profileImageType: selectedType,
-      profileImageUrl: selectedType === 'CUSTOM' ? customImageUrl : null,
-      customImageFile: selectedType === 'CUSTOM' ? customImageFile : null,
-    }),
-    [customImageFile, customImageUrl, selectedType],
-  );
-
-  const handleClose = () => {
-    onOpenChange(false);
-  };
-
-  const handleDialogOpenChange = (nextOpen: boolean) => {
-    if (!nextOpen && requireSubmitToClose && !allowCloseRef.current) {
-      return;
-    }
-
-    onOpenChange(nextOpen);
-  };
-
-  const handleCustomImageChange = (
-    event: React.ChangeEvent<HTMLInputElement>,
-  ) => {
-    const file = event.target.files?.[0];
-
-    if (!file) {
-      return;
-    }
-
-    if (objectUrlRef.current) {
-      URL.revokeObjectURL(objectUrlRef.current);
-    }
-
-    const nextObjectUrl = URL.createObjectURL(file);
-    objectUrlRef.current = nextObjectUrl;
-
-    setSelectedType('CUSTOM');
-    setCustomImageFile(file);
-    setCustomImageUrl(nextObjectUrl);
-  };
-
-  const handleSubmit = async () => {
-    try {
-      setIsSubmitting(true);
-      await onSubmit(selectedValue);
-      allowCloseRef.current = true;
-      onOpenChange(false);
-    } finally {
-      setIsSubmitting(false);
-      allowCloseRef.current = false;
-    }
-  };
+  const {
+    fileInputRef,
+    selectedType,
+    customImageUrl,
+    previewUrl,
+    isSubmitDisabled,
+    isSubmitting,
+    setSelectedType,
+    handleClose,
+    handleDialogOpenChange,
+    handleCustomImageChange,
+    handleSubmit,
+  } = useProfileImageEditDialog({
+    initialValue,
+    onOpenChange,
+    onSubmit,
+    requireSubmitToClose,
+  });
 
   return (
     <Dialog open={open} onOpenChange={handleDialogOpenChange}>
